refactor: split lexer.c
